Change ChunkDirection to ChunkSupport class
Might contain more enums or utility methods purely for chunks/pixels
