--- PixelProject/ChunkDirection.h
@@ -1,14 +1,0 @@
-#pragma once
-
-enum E_ChunkDirection : int
-{
-   North = 0,
-   NorthEast = 1,
-   East = 2,
-   SouthEast = 3,
-   South = 4,
-   SouthWest = 5,
-   West = 6,
-   NorthWest = 7,
-   DIR_COUNT = 8
-};
